Fix col_width calculation
Calculation was based only on length of largest item
in Header row of results, not all results.

Now it makes columns as large as the largest item
among all items within the results.  This fixes the small header
probelm, but now all columns are as large as the largest.
Now we need to calculate each column's width
separately.